+++ CHANGELOG.md
@@ -6,5 +6,6 @@ All notable changes in **python-transip** are documented below.
 - This `CHANGELOG.md` file to be able to list all notable changes for each version of **python-transip**.
 - The `transip.v6.objects.ApiTestService` service to allow calling the test resource to make sure everything is working.
 - The `transip.v6.objects.InvoiceItemService` service to allow listing all invoice items on a `transip.v6.objects.Invoice` object.
+- The `transip.mixins.ObjectUpdateMixin` mixin to allow calling `update()` on API object directly.
 
 [Unreleased]: https://github.com/roaldnefs/python-transip/compare/v0.3.0...HEAD